fuse works! it lists the fake file, now to RPC
The fuse app starts up, but the issue is that it's blocking, which sort of
makes sense. We only want to start fuse if we can establish a connection with
the remote machine. We can do this with the ping but then we need a few global
variables, which makes me uncomfortable. Let's see what happens.